Move display of articles to seperate component
