+++ src/components/App.js
@@ -1,8 +1,9 @@
 import ArticlesContextProvider from '../contexts/ArticlesContext';
 import { Routes, Route } from 'react-router-dom';
 import Header from './Header';
+import Search from './Search';
 import Sections from './Sections';
-import Articles from './Articles';
+import Main from './Main';
 import ArticleDetails from './ArticleDetails';
 import '../css/App.css';
 
@@ -11,12 +12,13 @@ const App = () => {
     <ArticlesContextProvider>
       <nav>
         <Header />
+        <Search />
         <Sections />
       </nav>
       <main>
         <Routes>
-          <Route path="/" element={<Articles />} />
-          <Route path="/:section" element={<Articles />} />
+          <Route path="/" element={<Main />} />
+          <Route path="/:section" element={<Main />} />
           <Route path="/:section/:title" element={<ArticleDetails />} />
         </Routes>
       </main>
