+++ src/css/Main.css
@@ -1,0 +1,3 @@
+.main-container {
+  
+}
